refactor: Start 제거
