+++ src/test/java/chess/model/board/state/StatusTest.java
@@ -20,7 +20,7 @@ class StatusTest {
     @DisplayName("calculateScores() 호출하면 첨수가 반환된다.")
     void calculateScores_whenCall_thenReturnScores() {
         // given
-        final GameState status = Start.from(STATUS);
+        final GameState status = ProgressState.from(STATUS);
 
         // when
         final Scores scores = status.calculateScores();
